# UCP/TOPO/BUG: Add memory distance only for relevant devices